+++ pytests/test_orphans_benefit.py
@@ -51,41 +51,3 @@ class TestOrphansBenefitForGrandparent(OrphansTests):
     def test_reasoning(self):
         self.assertTrue(self.is_conclusive)
         self.assertTrue(self.is_permitted)
-
-
-class TestOrphansBenefitForCarer(OrphansTests):
-
-    body = {
-        "applicant": {
-            "isPrincipalCarerForOneYearFromApplicationDate": True,
-            "Age": 21,
-            "isParent": False,
-            "isNZResident": True,
-            "normallyLivesInNZ": True
-        },
-        "child": {"isDependent": True},
-        "parents": {"areDeceasedMissingOrIncapableThroughDisability": True},
-    }
-
-    def test_reasoning(self):
-        self.assertTrue(self.is_conclusive)
-        self.assertTrue(self.is_permitted)
-
-
-class TestOrphansBenefitForChildForbidden(OrphansTests):
-
-    body = {
-        "applicant": {
-            "isPrincipalCarerForOneYearFromApplicationDate": True,
-            "Age": 17,
-            "isParent": False,
-            "isNZResident": True,
-            "normallyLivesInNZ": True
-        },
-        "child": {"isDependent": True},
-        "parents": {"areDeceasedMissingOrIncapableThroughDisability": True},
-    }
-
-    def test_reasoning(self):
-        self.assertTrue(self.is_conclusive)
-        self.assertTrue(self.is_forbidden)
